consertando várias coisas e primeiro teste com sucesso

+++ PBL1_AP2/src/modulo/Cardapio.java
@@ -2,21 +2,21 @@ package modulo;
 
 class Cardapio {
 	String decricao;
-	float valor;
+	double valor;
 	public String getDecricao() {
 		return decricao;
 	}
 	public void setDecricao(String decricao) {
 		this.decricao = decricao;
 	}
-	public float getValor() {
+	public double getValor() {
 		return valor;
 	}
 	public void setValor(float valor) {
 		this.valor = valor;
 	}
 	
-	public Cardapio(String des,float val) {
+	public Cardapio(String des,double val) {
 		this.decricao = des;
 		this.valor = val;
 	}
